Rename moreProjectsCard component to moreProjectCard
